feat: ele-4721 switch cdn to cloudfront distribution
BREAKING CHANGE: This will change the CDK class from
`cloudfront.CloudFrontWebDistribution` to `cloudfront.Distribution`
for the cdn-site-hosting-construct. This will create a slightly
different cfn template and so will deploy a new stack instead
invalidating the old one. This will then break as the dns
as it will be the same.